Pin dependencies to specific versions
The current requirements.txt does not specify any version for
dependencies which results in the current latest version available
frmo PyPi being pulled in.

In order to make the build/setup process reproducible, the dependencies
should be pinned to specific versions. This can easily be achieved
by using the 'pip-compile' command of the 'pip-tools' module which
takes dependency declarations from the 'requirements.in' file and
compiles these into a 'requirements.txt' file containing specific
versions of the dependencies.

Also replaced the dependency on j1939 with can-j1939. The GitHub
repo of the former has not seen any updates for five years while
the repo of the latter seems to be under active development.
latter seems to be a fork of the former and